Vector-index-api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
qdrant-client==1.12.1
httpx==0.27.2
Pillow==10.4.0
requests==2.32.3
open_clip_torch==2.24.0
numpy==1.26.4